Crate quill_source_file

Structs

ErrorEmitter

Prints error and warning messages, outputting the relevant lines of source code from the input files.

Module

A tree of source files and other modules.

PackageFileSystem

Represents the file structure of an entire package on disk. A package is a folder that produces some kind of compiled output file, which could be a binary, a static library, a shared object, a WASM module, or something else of that kind.

SourceFile

A single file of source code.

Enums

SourceFileLoadError

If a source file’s contents could not be loaded, why was this?

Functions

find_all_source_files